requirements.txt
# ============================================================================
# OpenVINO GenAI API Server - All Dependencies
# Install everything: pip install -r requirements.txt
# ============================================================================

# Core API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# OpenVINO & AI
openvino>=2024.0.0
openvino-genai>=2024.0.0

# Data Processing
numpy>=1.24.0
Pillow>=10.0.0

# HTTP & Networking
requests>=2.31.0
websockets>=12.0  # For WebSocket Realtime API

# Client SDK (for examples/testing)
openai>=1.0.0

# Document Processing (RAG)
PyPDF2>=3.0.0  # PDF extraction
python-docx>=1.0.0  # Word document extraction

# Audio Processing (TTS/STT)
librosa>=0.10.0  # Audio loading for Whisper
soundfile>=0.12.0  # Audio file I/O
sounddevice>=0.4.6  # Microphone/speaker (voice chat)
pydub>=0.25.0  # Audio format conversion (optional)

# Model Conversion Tools (optional - for setup only)
# Uncomment if you need to convert models:
hf_xet
huggingface_hub[hf_xet]
git+https://github.com/huggingface/optimum-intel.git
nncf